Move gcwWrite to util.go, use it generally
