+++ kubetest/kops.go
@@ -516,28 +516,3 @@ func (k kops) Publish() error {
 		return gcsWrite(k.kopsVersion, []byte(k.kopsPublish))
 	})
 }
-
-// gcsWrite uploads contents to the dest location in GCS.
-// It currently shells out to gsutil, but this could change in future.
-func gcsWrite(dest string, contents []byte) error {
-	f, err := ioutil.TempFile("", "")
-	if err != nil {
-		return fmt.Errorf("error creating temp file: %v", err)
-	}
-
-	defer func() {
-		if err := os.Remove(f.Name()); err != nil {
-			log.Printf("error removing temp file: %v", err)
-		}
-	}()
-
-	if _, err := f.Write(contents); err != nil {
-		return fmt.Errorf("error writing temp file: %v", err)
-	}
-
-	if err := f.Close(); err != nil {
-		return fmt.Errorf("error closing temp file: %v", err)
-	}
-
-	return finishRunning(exec.Command("gsutil", "cp", f.Name(), dest))
-}
